pyproject.toml
[project]
name = "oberon-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mcp[cli]>=1.9.4",
    "nest-asyncio>=1.6.0",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
]
